src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.10.0.84
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.0
python-dotenv==1.0.1
huggingface_hub==0.36.0
